Renamed `include` folder to `gsl` to enable the use of consistent #include directives regardless if the library is copied-in, installed or included from the development environment.
