+++ gsl/gsl.h
@@ -19,12 +19,12 @@
 #ifndef GSL_GSL_H
 #define GSL_GSL_H
 
-#include "gsl_assert.h"  // Ensures/Expects
-#include "gsl_util.h"    // finally()/narrow()/narrow_cast()...
-#include "multi_span.h"  // multi_span, strided_span...
-#include "span.h"        // span
-#include "string_span.h" // zstring, string_span, zstring_builder...
 #include <memory>
+#include "../gsl/gsl_assert.h"  // Ensures/Expects
+#include "../gsl/gsl_util.h"    // finally()/narrow()/narrow_cast()...
+#include "../gsl/multi_span.h"  // multi_span, strided_span...
+#include "../gsl/span.h"        // span
+#include "../gsl/string_span.h" // zstring, string_span, zstring_builder...
 
 #ifdef _MSC_VER
