+++ gsl/multi_span.h
@@ -19,9 +19,6 @@
 #ifndef GSL_MULTI_SPAN_H
 #define GSL_MULTI_SPAN_H
 
-#include "gsl_assert.h"
-#include "gsl_byte.h"
-#include "gsl_util.h"
 #include <algorithm>
 #include <array>
 #include <cassert>
@@ -35,6 +32,9 @@
 #include <stdexcept>
 #include <type_traits>
 #include <utility>
+#include "../gsl/gsl_assert.h"
+#include "../gsl/gsl_byte.h"
+#include "../gsl/gsl_util.h"
 
 #ifdef _MSC_VER
