+++ gsl/span.h
@@ -20,15 +20,15 @@
 #ifndef GSL_SPAN_H
 #define GSL_SPAN_H
 
-#include "gsl_assert.h"
-#include "gsl_byte.h"
-#include "gsl_util.h"
 #include <array>
 #include <iterator>
 #include <limits>
 #include <stdexcept>
 #include <type_traits>
 #include <utility>
+#include "../gsl/gsl_assert.h"
+#include "../gsl/gsl_byte.h"
+#include "../gsl/gsl_util.h"
 
 #ifdef _MSC_VER
