+++ tests/assertion_tests.cpp
@@ -15,7 +15,8 @@
 ///////////////////////////////////////////////////////////////////////////////
 
 #include <UnitTest++/UnitTest++.h> 
-#include <gsl.h>
+
+#include "../gsl/gsl.h"
 
 using namespace gsl;
