+++ tests/byte_tests.cpp
@@ -15,14 +15,13 @@
 ///////////////////////////////////////////////////////////////////////////////
 
 #include <UnitTest++/UnitTest++.h>
-#include <gsl_byte.h>
-
 #include <iostream>
 #include <list>
 #include <map>
 #include <memory>
 #include <string>
 #include <vector>
+#include "../gsl/gsl_byte.h"
 
 using namespace std;
 using namespace gsl;
