+++ tests/string_span_tests.cpp
@@ -16,8 +16,8 @@
 
 #include <UnitTest++/UnitTest++.h>
 #include <cstdlib>
-#include <string_span.h>
 #include <vector>
+#include "../gsl/string_span.h"
 
 using namespace std;
 using namespace gsl;
